chore: fleet do not play well with ELB
With ElasticLoadBalancingV2 we need to define a TargetGroup
and our instances must be part of this TargetGroup.

When we create our instances with EC2Fleet, we have no way
to declare that we want them to be attached to some TargetGroup.
The only way to attach these instances would be to have some
lambda reacting to their creation and dynamically attaching
them to some TargetGroup... that does not look like something
we want to do and maintain.

Since we only want two instances, let's switch to a more
static approach that can then be integrated with TargetGroup.

See:
* https://docs.aws.amazon.com/AWSCloudFormation/latest/TemplateReference/aws-resource-elasticloadbalancingv2-targetgroup.html
* https://docs.aws.amazon.com/AWSCloudFormation/latest/TemplateReference/aws-resource-ec2-ec2fleet.html